mesh: Update AppKeys on transition to Phase 0
At the end of the mesh Key Refresh procedure when a subnet
transitions to Phase 0, local AppKeys that were updated were
not updating until the bluetooth-meshd daemon was restarted.

This patch iterates the AppKeys at the end of mesh Key Refresh
when the subnet transitions to Phase 0, setting the new state
of each updated AppKey.

********* v2 ************

This version incorporates recommendations from Brian Gix
creating a public function appkey_finalize(); and adding
new APP_AID_INVALID constant to replace the use of 0xFF
as an indication that the new_key_aid field is invalid.